# openapi.yaml
openapi: 3.0.3
info:
  title: FinMate API
  description: |
    API Contract for FinMate, a Personal Finance & Lifestyle Companion.
    Provides endpoints for Auth, Users, Groups, Expenses, Settlements, Notes, Goals, and Import/Export.

    ### Versioning Strategy
    URL-based versioning is used. All paths start with `/api/v1`.

    ### Pagination & Filtering
    Paginated GET endpoints support standard query parameters:
    * `page`: 1-based page index (default: `1`)
    * `limit`: Page size (default: `20`, max: `100`)
    * `cursor`: Cursor string (for cursor-based pagination on rapidly updated lists like Expenses and Audit Logs)
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local Dev Server
paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: securePassword123!
                displayName:
                  type: string
                  maxLength: 120
                  example: John Doe
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /auth/login:
    post:
      summary: Login user and retrieve JWT tokens
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123!
                deviceId:
                  type: string
                  example: 'web-client-chrome-1234'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - refreshToken
                  - user
                properties:
                  accessToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
                  refreshToken:
                    type: string
                    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      summary: Refresh access token using a valid refresh token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - refreshToken
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      summary: Invalidate user session
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/2fa/enable:
    post:
      summary: Initiate 2FA TOTP setup
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 2FA initialization details returned
          content:
            application/json:
              schema:
                type: object
                required:
                  - secret
                  - qrCodeUrl
                properties:
                  secret:
                    type: string
                    example: 'KVKFKRCSN5RHK33O'
                  qrCodeUrl:
                    type: string
                    example: 'otpauth://totp/FinMate:user@example.com?secret=KVKFKRCSN5RHK33O&issuer=FinMate'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/2fa/verify:
    post:
      summary: Verify and complete 2FA setup
      tags:
        - Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  pattern: '^[0-9]{6}$'
                  example: '123456'
      responses:
        '200':
          description: 2FA enabled successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/2fa/disable:
    post:
      summary: Disable 2FA for the account
      tags:
        - Auth
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  pattern: '^[0-9]{6}$'
                  example: '123456'
      responses:
        '204':
          description: 2FA disabled successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /expenses:
    post:
      summary: Create a personal or group expense
      tags:
        - Expenses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - amountTotal
                - currency
                - category
                - paidByUserId
                - expenseDate
                - splits
              properties:
                title:
                  type: string
                  maxLength: 1000
                  example: 'Dinner at Taj'
                description:
                  type: string
                  example: 'Celebration dinner'
                amountTotal:
                  type: number
                  format: float
                  example: 4500.00
                currency:
                  type: string
                  maxLength: 3
                  example: 'INR'
                category:
                  type: string
                  example: 'Food'
                paidByUserId:
                  type: string
                  format: uuid
                groupId:
                  type: string
                  format: uuid
                  nullable: true
                expenseDate:
                  type: string
                  format: date
                  example: '2026-06-09'
                status:
                  type: string
                  enum: [draft, posted, void]
                  default: posted
                splits:
                  type: array
                  items:
                    $ref: '#/components/schemas/ExpenseSplitInput'
                attachmentKeys:
                  type: array
                  items:
                    type: string
                  example: ['receipts/20260609_taj.jpg']
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    get:
      summary: List expenses with pagination and filtering
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: cursor
          in: query
          schema:
            type: string
          required: false
        - name: groupId
          in: query
          schema:
            type: string
            format: uuid
          required: false
        - name: category
          in: query
          schema:
            type: string
          required: false
        - name: startDate
          in: query
          schema:
            type: string
            format: date
          required: false
        - name: endDate
          in: query
          schema:
            type: string
            format: date
          required: false
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExpenseResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /expenses/{id}:
    get:
      summary: Retrieve specific expense details
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Expense details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      summary: Update expense details and splits
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - version
              properties:
                title:
                  type: string
                  maxLength: 1000
                description:
                  type: string
                amountTotal:
                  type: number
                currency:
                  type: string
                  maxLength: 3
                category:
                  type: string
                paidByUserId:
                  type: string
                  format: uuid
                expenseDate:
                  type: string
                  format: date
                status:
                  type: string
                  enum: [draft, posted, void]
                splits:
                  type: array
                  items:
                    $ref: '#/components/schemas/ExpenseSplitInput'
                attachmentKeys:
                  type: array
                  items:
                    type: string
                version:
                  type: integer
                  example: 1
      responses:
        '200':
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '412':
          $ref: '#/components/responses/PreconditionFailedError'
    delete:
      summary: Void or remove an expense
      tags:
        - Expenses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Expense voided or deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of elements per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 400
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/auth/register'
            errorCode: 'VAL_INVALID_INPUT'
            message: 'Input validation failed'
            details:
              - field: 'email'
                issue: 'must be a valid email address'
            retryable: false
    UnauthorizedError:
      description: Authentication failed or token missing/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/users/me'
            errorCode: 'AUTH_TOKEN_EXPIRED'
            message: 'JWT access token has expired'
            retryable: false
    ForbiddenError:
      description: Permissions not met for resource action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 403
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/groups/d3b07384-d113-49c3-a558-f54248880bf6'
            errorCode: 'RES_FORBIDDEN'
            message: 'You do not have access to view this group.'
            retryable: false
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/expenses/f81d4fae-7dec-11d0-a765-00a0c91e6bf6'
            errorCode: 'RES_NOT_FOUND'
            message: 'Expense with ID f81d4fae-7dec-11d0-a765-00a0c91e6bf6 does not exist.'
            retryable: false
    ConflictError:
      description: Resource constraint or duplication conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 409
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/auth/register'
            errorCode: 'RES_ALREADY_EXISTS'
            message: 'A user with email user@example.com already exists.'
            retryable: false
    PreconditionFailedError:
      description: State conflict or optimistic lock version mismatch
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 412
            timestamp: '2026-06-09T17:15:00.000Z'
            path: '/api/v1/expenses/f81d4fae-7dec-11d0-a765-00a0c91e6bf6'
            errorCode: 'CON_VERSION_CONFLICT'
            message: 'The resource version has changed. Please fetch the latest state and retry.'
            retryable: true
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - timestamp
        - path
        - errorCode
        - message
        - retryable
      properties:
        statusCode:
          type: integer
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        errorCode:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required:
              - field
              - issue
            properties:
              field:
                type: string
              issue:
                type: string
        retryable:
          type: boolean
    PaginatedResponse:
      type: object
      required:
        - data
        - meta
        - links
      properties:
        meta:
          type: object
          required:
            - totalItems
            - itemCount
            - itemsPerPage
            - totalPages
            - currentPage
          properties:
            totalItems:
              type: integer
            itemCount:
              type: integer
            itemsPerPage:
              type: integer
            totalPages:
              type: integer
            currentPage:
              type: integer
        links:
          type: object
          required:
            - first
            - last
          properties:
            first:
              type: string
            previous:
              type: string
              nullable: true
            next:
              type: string
              nullable: true
            last:
              type: string
    UserResponse:
      type: object
      required:
        - id
        - email
        - displayName
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, disabled, invited]
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ExpenseResponse:
      type: object
      required:
        - id
        - title
        - description
        - amountTotal
        - currency
        - category
        - paidByUserId
        - ownerUserId
        - groupId
        - expenseDate
        - status
        - splits
        - attachments
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        amountTotal:
          type: number
        currency:
          type: string
        category:
          type: string
        paidByUserId:
          type: string
          format: uuid
        ownerUserId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
          nullable: true
        expenseDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, posted, void]
        splits:
          type: array
          items:
            $ref: '#/components/schemas/ExpenseSplit'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        version:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ExpenseSplit:
      type: object
      required:
        - id
        - expenseId
        - splitType
        - shareValue
        - amountOwed
        - isSettled
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        expenseId:
          type: string
          format: uuid
        participantUserId:
          type: string
          format: uuid
          nullable: true
        participantGroupMemberId:
          type: string
          format: uuid
          nullable: true
        splitType:
          type: string
          enum: [equal, fixed, percent, share]
        shareValue:
          type: number
        amountOwed:
          type: number
        isSettled:
          type: boolean
        settledAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ExpenseSplitInput:
      type: object
      required:
        - splitType
        - shareValue
      properties:
        participantUserId:
          type: string
          format: uuid
          nullable: true
        participantGroupMemberId:
          type: string
          format: uuid
          nullable: true
        splitType:
          type: string
          enum: [equal, fixed, percent, share]
        shareValue:
          type: number
    Attachment:
      type: object
      required:
        - id
        - uploaderUserId
        - storageKey
        - originalName
        - mimeType
        - sizeBytes
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        uploaderUserId:
          type: string
          format: uuid
        expenseId:
          type: string
          format: uuid
          nullable: true
        noteId:
          type: string
          format: uuid
          nullable: true
        goalId:
          type: string
          format: uuid
          nullable: true
        groupId:
          type: string
          format: uuid
          nullable: true
        storageKey:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
        checksumSha256:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...